pyproject.toml
[project]
name = "latex-beamer-converter"
version = "0.1.0"
description = "Intelligent agent that converts research papers (LaTeX) into professional Beamer presentations"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["latex", "beamer", "presentation", "converter", "academic", "slides"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Markup :: LaTeX",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "pydantic>=2.6",
  "rich>=13.7",
  "typer>=0.12",
]

[project.urls]
Homepage = "https://github.com/yourusername/AR"
Documentation = "https://github.com/yourusername/AR#readme"
Repository = "https://github.com/yourusername/AR.git"
Issues = "https://github.com/yourusername/AR/issues"

[project.scripts]
agent = "agent.cli:app"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
only-include = ["src/agent"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
packages = ["src/agent"]
